server/requirements.txt
Django==3.2
pillow
face-recognition==1.3